FirstPerson View WIP
Added a character and created a CharacterManager script. For the moment we can move and look around as the character. The character is not yet compatible with the timeline. Problem with character camera which lose correct orientation over time. Camera swap (main / sub or sub / sub) is done.
Should we be able to record from first person view ? (For the moment we say no).